% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.azk.R
\name{read.azk}
\alias{read.azk}
\title{Read DMDX item file}
\usage{
read.azk(file, acc = FALSE, item.order = FALSE, read.max = 20000,
  warn = FALSE)
}
\arguments{
\item{file}{A name of the file}

\item{acc}{logical. If \code{TURE}, negative reaction time
(DMDX's way of marking the trials with incorrect response) will
be converted to positive, and then new column containing response accuracy information
will be added.(column named \code{ACC}); Default to \code{FALSE}}

\item{item.order}{logical. If \code{TRUE}, column containing item order will be produced. Default to \code{FALSE}}

\item{read.max}{maximum number of lines to read at a time during initial parsing of the data file.}

\item{warn}{logical. If \code{FALSE} (Default), messages form internally used
functions (e.g., readLines, scan) will be muted.}
}
\value{
A data.frame containing following columns:
\itemize{
 \item{sbj, subject number within the data file}
 \item{date, data colletion dat}
 \item{time, data collection time}
 \item{machine, name of the data collection machine}
 \item{refresh, reflecsh rate used}
 \item{id, ID assigned to the subject}
 \item{RT, response latency}
} and optionally,
\itemize{
 \item{ACC, accuracy of responses}
 \item{item.order, order of the items}
}
}
\description{
Read DMDX item file in either RTF format or ASCII format, and convert it into a list contaning the header information and data frame
}

