% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dmdx.R
\name{read.dmdx}
\alias{read.dmdx}
\title{Read DMDX item file}
\usage{
read.dmdx(file, colnames, CorrectAnswers = FALSE, warn = FALSE,
  separator = FALSE, encoding = "unknown")
}
\arguments{
\item{file}{A name of file}

\item{colnames}{names of columns for the data.frame to be returned.}

\item{CorrectAnswers}{logical. If \code{TURE}, the signs (+, -, ^, =) will
be removed from the item number and converted into a new column 'Correct Answers
with the corresponding values ("Yes", "No", \code{NA}, and "Any" respectively).
Default to \code{FALSE}}

\item{warn}{logical. If \code{FALSE} (Default), messages form internally used
functions (e.g., readLines) will be muted.}

\item{separator}{logical. If \code{FALSE} (Default), frame separator '/' will be omitted from the output data.frame.}

\item{encoding}{encoding to be used.}
}
\value{
A list containing:
\itemize{
  \item{header (A list of header items. Each <> are treated as separate items)}
  \item{itemdat (A data frame with each column corresponding to item number,
  item from each frame marked by the divider /, and timing and other indicators.)}
}
}
\description{
Read DMDX item file in either RTF format or ASCII format, and convert it into a list contaning the header information and data frame
}

