% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtf2textvec.R
\name{rtf2textvec}
\alias{rtf2textvec}
\title{Convert RTF file into a plain-text charactor vector}
\usage{
rtf2textvec(file, warn = FALSE)
}
\arguments{
\item{file}{A name of the file}

\item{warn}{logical. If \code{FALSE} (Default), messages form internally used
functions (e.g., readLines) will be muted.}
}
\value{
A character vector with each element containing stripped out texts form each line in the file.
}
\description{
Convert RTF file format into a charactor vector, with each element containing stripped out texts form each line in the file.
}

